pyproject.toml
[project]
name = "grumpyclaw"
version = "0.1.0"
description = "Personal AI agent: SQLite + Markdown memory, heartbeat, terminal."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastembed>=0.3.0",
    "fastapi>=0.115.0",
    "fastrtc>=0.0.34",
    "numpy>=1.24.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "reachy-mini>=1.3.1",
    "reachy-mini-dances-library",
    "reachy-mini-toolbox",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
grumpyadmin-api = "api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grumpyclaw", "src/grumpyreachy", "api"]
